wpcom-vip adapter: lowercasing meta queries containing arrays
MySQL on the WordPress.com platform is case insensitive and we're mapping all meta query comparisons to `raw_lc` and we are already normalising strings to lowercase.

But as WP_Meta_Query allows developers to pass arrays of values to some fields (eg.: meta_value), we need to normalise even those arrays, not only strings.

This is related to #11 and was mentioned in Zendesk #51832