More compatibility with WP_Query, when sorting using meta-query keys
WP_Query supports specifying meta-queries such as these, where $args are the
arguments to WP_Query() :

    $args['meta_query'][]['pv_date'] = array(
        'key' => 'pv_date',
        'type' => 'NUMERIC',
        'compare' => 'EXISTS',
    );

    $args['meta_query'][]['pv_count'] = array(
        'key' => 'pv_count',
        'type' => 'NUMERIC',
        'compare' => 'EXISTS',
    );

Currently these are processed correctly by es-wp-query, but associated
ordering clauses are not, for instance:

    $args["orderby"] => Array(
        "pv_date" => "desc",
    	"pv_count" => "desc"
    );

This is because sorting-arguments are filtered, so unrecognized ones are left
out. However, that does not include meta-queries that are specified as shown
above.

The patch will ensure processing of these meta-queries, and allow sorting to
take place.