Image Search Enhancements.
ES_WP_Query currently does a prefix search on the post_mime_type field. To perform this type of query the raw field needs to be utilizied.

Currently no support exists for the "inherit" post_status search. This is the base post_status used when doing image searches. This patch sets the post_type equals to all searchable post_status (same as any).