fix: jetpack search initializer blog id
The expected value for the classic search initializer is the site's wpcom ID, not its multisite blog ID.

The issue only appears when Jetpack's Instant Search module is enabled, which prevents the Classic_Search singleton from being first initialized. When initializing through the ES_WP_Query it's misconfigured and fails to perform requests.